refactor: restructure project to move v1 models and services to internal package
