+++ exchange/policy-decision-point/internal/db/database_test.go
@@ -1,4 +1,4 @@
-package v1
+package db
 
 import (
 	"testing"
