+++ exchange/policy-decision-point/internal/handler/handler_test.go
@@ -1,4 +1,4 @@
-package v1
+package handler
 
 import (
 	"bytes"
@@ -7,8 +7,8 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/OpenNDX/openndx-core/exchange/policy-decision-point/v1/models"
-	"github.com/OpenNDX/openndx-core/exchange/policy-decision-point/v1/testhelpers"
+	"github.com/OpenNDX/openndx-core/exchange/policy-decision-point/internal/models"
+	"github.com/OpenNDX/openndx-core/exchange/policy-decision-point/internal/testhelpers"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gorm.io/gorm"
